+++ src/main/kotlin/hu/netsurf/erp/warehouse/repository/SupplierRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface SupplierRepository : JpaRepository<Supplier, Int>
+interface SupplierRepository : JpaRepository<Supplier, Long>
