+++ src/test/kotlin/hu/netsurf/erp/usermanagement/testobject/DeleteProfileInputTestObject.kt
@@ -6,6 +6,6 @@ class DeleteProfileInputTestObject {
     companion object {
         fun input1(): DeleteProfileInput = build()
 
-        private fun build(userId: Int = 1): DeleteProfileInput = DeleteProfileInput(userId = userId)
+        private fun build(userId: Long = 1): DeleteProfileInput = DeleteProfileInput(userId = userId)
     }
 }
